Rework sprite systems yet again.
Now there is only a sprite, with the stage properties
as regular variables in main.

Now the program class calls threads based on an index
property in the thread in reference to a new vector of objects.
In the future this should be expanded so that `None` means the stage.