Fix an issue with modifying the stage.
Now the stage is (again!) a struct. This time, instead of trying to pass
a reference to it with the sprite(that caused lifetime problems), we
pass a cloned version of the stage along with the sprite. Then, when we
finish ticking the thread, we set the "master" stage to the intermediate
stage.  This means more clones, but it works.